ocr/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
numpy
Pillow
pydantic
# surya-ocr
